requirements.txt
google-ai-generativelanguage
google-api-core
google-auth
google-generativeai
pyppeteer